requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
SQLAlchemy==2.0.31
serpy==0.3.1
djangorestframework-simplejwt==5.3.1
